Add SGX C types.
