+++ src/sgx.rs
@@ -1,0 +1,5 @@
+//! SGX C types definition
+
+pub type c_char = i8;
+pub type c_long = i64;
+pub type c_ulong = u64;
